[SPARKC-427] Updated based on PR comments
* Moved annotationForFieldName to JavaBeanColumnMapper
* Added extra test to map Cassandra fields directly to POJO
