+++ spark-cassandra-connector/src/main/scala/com/datastax/spark/connector/mapper/JavaBeanColumnMapper.scala
@@ -3,6 +3,7 @@ package com.datastax.spark.connector.mapper
 import java.lang.reflect.Method
 
 import com.datastax.driver.core.ProtocolVersion
+import com.datastax.driver.mapping.annotations.{Column, Field}
 import com.datastax.spark.connector.ColumnRef
 import com.datastax.spark.connector.cql.TableDef
 
@@ -32,6 +33,24 @@ class JavaBeanColumnMapper[T : ClassTag](columnNameOverride: Map[String, String]
     SetterRegex.findFirstMatchIn(method.getName).isDefined &&
     method.getParameterTypes.size == 1 &&
     method.getReturnType == Void.TYPE
+
+  private def annotationForFieldName(fieldName: String): String = {
+    // POJO is either a table or an UDT
+    // We have to cover both cases
+    try {
+      val f = cls.getField(fieldName)
+
+      if (f.isAnnotationPresent(classOf[Column])) {
+        f.getAnnotation(classOf[Column]).name()
+      } else if (f.isAnnotationPresent(classOf[Field])) {
+        f.getAnnotation(classOf[Field]).name()
+      } else {
+        ""
+      }
+    } catch {
+      case e: NoSuchFieldException => { "" }
+    }
+  }
 
   private def resolve(name: String, columns: Map[String, ColumnRef]): Option[ColumnRef] = {
     val overridenName = columnNameOverride.getOrElse(name, name)
