+++ spark-cassandra-connector/src/main/scala/com/datastax/spark/connector/mapper/ReflectionColumnMapper.scala
@@ -2,7 +2,6 @@ package com.datastax.spark.connector.mapper
 
 import java.lang.reflect.{Constructor, Method}
 
-import com.datastax.driver.mapping.annotations.{Column, Field}
 import com.datastax.spark.connector.ColumnRef
 import com.datastax.spark.connector.cql.StructDef
 import com.datastax.spark.connector.rdd.reader.AnyObjectFactory
@@ -29,32 +28,12 @@ abstract class ReflectionColumnMapper[T : ClassTag] extends ColumnMapper[T] {
       paramName: String,
       columns: Map[String, ColumnRef]): Option[ColumnRef]
 
-  protected def annotationForFieldName(
-      fieldName: String): String = {
-    // POJO is either a table or an UDT
-    // We have to cover both cases
-    try {
-      val f = cls.getField(fieldName)
-
-      if (f.isAnnotationPresent(classOf[Column])) {
-        f.getAnnotation(classOf[Column]).name()
-      } else if (f.isAnnotationPresent(classOf[Field])) {
-        f.getAnnotation(classOf[Field]).name()
-      } else {
-        ""
-      }
-    } catch {
-      case e: NoSuchFieldException => { "" }
-    }
-
-  }
-
   protected def allowsNull: Boolean
 
   private def columnRefByAliasName(selectedColumns: IndexedSeq[ColumnRef]): Map[String, ColumnRef] =
     (for (c <- selectedColumns) yield (c.selectedAs, c)).toMap
 
-  private val cls = SparkReflectionLock.synchronized { implicitly[ClassTag[T]].runtimeClass }
+  protected val cls = SparkReflectionLock.synchronized { implicitly[ClassTag[T]].runtimeClass }
 
   override def columnMapForReading(
       struct: StructDef,
